blog_backend/pyproject.toml
[project]
name = "blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "cryptography>=46.0.5",
    "fastapi>=0.128.8",
    "openai>=2.26.0",
    "passlib>=1.7.4",
    "playwright>=1.58.0",
    "pymysql>=1.1.2",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=7.0.1",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.39.0",
]
